refactored modules into utils package
